Move unmanaged-to-managed tracking types to a separate file to help encapsulate it.
