+++ src/libraries/System.Runtime.InteropServices/gen/Microsoft.Interop.SourceGeneration/Marshalling/AttributedMarshallingModelGeneratorFactory.cs
@@ -248,7 +248,7 @@ namespace Microsoft.Interop
 
                 if (freeStrategy == FreeStrategy.FreeOriginal)
                 {
-                    marshallingStrategy = new StatelessUnmanagedToManagedOwnershipTracking(marshallingStrategy);
+                    marshallingStrategy = new UnmanagedToManagedOwnershipTrackingStrategy(marshallingStrategy);
                 }
 
                 if (freeStrategy != FreeStrategy.NoFree && marshallerData.Shape.HasFlag(MarshallerShape.Free))
@@ -258,7 +258,7 @@ namespace Microsoft.Interop
 
                 if (freeStrategy == FreeStrategy.FreeOriginal)
                 {
-                    marshallingStrategy = new FreeOwnedOriginalValueMarshalling(marshallingStrategy);
+                    marshallingStrategy = new CleanupOwnedOriginalValueMarshalling(marshallingStrategy);
                 }
             }
 
@@ -331,14 +331,14 @@ namespace Microsoft.Interop
 
                 if (freeStrategy == FreeStrategy.FreeOriginal)
                 {
-                    marshallingStrategy = new StatelessUnmanagedToManagedOwnershipTracking(marshallingStrategy);
+                    marshallingStrategy = new UnmanagedToManagedOwnershipTrackingStrategy(marshallingStrategy);
                 }
 
                 marshallingStrategy = new StatefulLinearCollectionMarshalling(marshallingStrategy, marshallerData.Shape, numElementsExpression, elementsMarshalling, freeStrategy != FreeStrategy.NoFree);
 
                 if (freeStrategy == FreeStrategy.FreeOriginal)
                 {
-                    marshallingStrategy = new FreeOwnedOriginalValueMarshalling(marshallingStrategy);
+                    marshallingStrategy = new CleanupOwnedOriginalValueMarshalling(marshallingStrategy);
                 }
 
                 if (marshallerData.Shape.HasFlag(MarshallerShape.Free))
@@ -355,7 +355,7 @@ namespace Microsoft.Interop
                 IElementsMarshallingCollectionSource collectionSource = new StatelessLinearCollectionSource(marshallerTypeSyntax);
                 if (freeStrategy == FreeStrategy.FreeOriginal)
                 {
-                    marshallingStrategy = new StatelessUnmanagedToManagedOwnershipTracking(marshallingStrategy);
+                    marshallingStrategy = new UnmanagedToManagedOwnershipTrackingStrategy(marshallingStrategy);
                 }
 
                 IElementsMarshalling elementsMarshalling = CreateElementsMarshalling(marshallerData, elementInfo, elementMarshaller, unmanagedElementType, collectionSource);
@@ -378,7 +378,7 @@ namespace Microsoft.Interop
 
                 if (freeStrategy == FreeStrategy.FreeOriginal)
                 {
-                    marshallingStrategy = new FreeOwnedOriginalValueMarshalling(marshallingStrategy);
+                    marshallingStrategy = new CleanupOwnedOriginalValueMarshalling(marshallingStrategy);
                 }
             }
